Add files via upload
this file is for the machine learning model that will actually analyze the handwriting. And this will be concatenated with a webapp(by other contributors) in future to form its full setup. This model is incomplete right now and will be completed as soon as possible and will we refurbished with CNNs.